Move the missing_tracks and missing_name tests from pytrainer/lib/test_gpx.py to pytrainer/test/lib/test_gpx.py
